-powSpecNu(), grid.x as linspace ++
There was an issue with Grid.x, the periodicity was violated.
I used np.linspace and enforced periodicity, but then I lost the
explicit 0. grid point (since J = 2N+1).
That is why I removed the numerical power spectrum method and the
uncorrelated graph.

Also the discrete Fourier transform does not order in term of
wavenumber per se, because it is a decomposition on sines and
cosines.
It was confusing to show the numerical spectrum on the wavenumber
axis...